refactor: mock 패키지 repository 디렉토리 생성
