+++ aics-admin/src/testFixtures/java/about/application/AboutAdminFacadeTest.java
@@ -21,7 +21,7 @@ import kgu.developers.domain.about.domain.MainCategory;
 import kgu.developers.domain.about.domain.SubCategory;
 import kgu.developers.domain.about.exception.AboutNotFoundException;
 import kgu.developers.domain.about.exception.CategoryNotMatchException;
-import mock.FakeAboutRepository;
+import mock.repository.FakeAboutRepository;
 
 public class AboutAdminFacadeTest {
 	private AboutAdminFacade aboutAdminFacade;
