+++ aics-api/src/testFixtures/java/about/application/AboutFacadeTest.java
@@ -19,7 +19,7 @@ import kgu.developers.domain.about.domain.MainCategory;
 import kgu.developers.domain.about.domain.SubCategory;
 import kgu.developers.domain.about.exception.AboutNotFoundException;
 import kgu.developers.domain.about.exception.CategoryNotMatchException;
-import mock.FakeAboutRepository;
+import mock.repository.FakeAboutRepository;
 
 public class AboutFacadeTest {
 	private AboutFacade aboutFacade;
