+++ aics-api/src/testFixtures/java/club/application/ClubFacadeTest.java
@@ -10,7 +10,7 @@ import kgu.developers.api.club.application.ClubFacade;
 import kgu.developers.api.club.presentation.response.ClubListResponse;
 import kgu.developers.domain.club.application.query.ClubQueryService;
 import kgu.developers.domain.club.domain.Club;
-import mock.FakeClubRepository;
+import mock.repository.FakeClubRepository;
 
 public class ClubFacadeTest {
 	private ClubFacade clubFacade;
