+++ aics-api/src/testFixtures/java/post/application/PostFacadeTest.java
@@ -28,8 +28,8 @@ import kgu.developers.domain.post.domain.Category;
 import kgu.developers.domain.post.domain.Post;
 import kgu.developers.domain.user.application.query.UserQueryService;
 import kgu.developers.domain.user.domain.User;
-import mock.FakePostRepository;
-import mock.FakeUserRepository;
+import mock.repository.FakePostRepository;
+import mock.repository.FakeUserRepository;
 
 public class PostFacadeTest {
 	private PostFacade postFacade;
