+++ aics-domain/src/testFixtures/java/comment/application/CommentCommandServiceTest.java
@@ -19,9 +19,9 @@ import kgu.developers.domain.post.application.query.PostQueryService;
 import kgu.developers.domain.post.domain.Post;
 import kgu.developers.domain.user.application.query.UserQueryService;
 import kgu.developers.domain.user.domain.User;
-import mock.FakeCommentRepository;
-import mock.FakePostRepository;
-import mock.FakeUserRepository;
+import mock.repository.FakeCommentRepository;
+import mock.repository.FakePostRepository;
+import mock.repository.FakeUserRepository;
 
 public class CommentCommandServiceTest {
 	private CommentCommandService commentCommandService;
