+++ aics-domain/src/testFixtures/java/mock/FakeTestContainer.java
@@ -35,6 +35,14 @@ import kgu.developers.domain.user.application.command.UserCommandService;
 import kgu.developers.domain.user.application.query.UserQueryService;
 import kgu.developers.domain.user.domain.User;
 import kgu.developers.domain.user.domain.UserRepository;
+import mock.repository.FakeAboutRepository;
+import mock.repository.FakeClubRepository;
+import mock.repository.FakeCommentRepository;
+import mock.repository.FakeLabRepository;
+import mock.repository.FakePostRepository;
+import mock.repository.FakeProfessorRepository;
+import mock.repository.FakeRefreshTokenRepository;
+import mock.repository.FakeUserRepository;
 
 public class FakeTestContainer {
 	private final Map<Class<?>, Supplier<?>> suppliers = new HashMap<>();
